requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==1.0.0
aspy.yaml==1.3.0
astroid==2.3.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.2.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.7.0
black==19.10b0
brotlipy==0.7.0
cchardet==2.1.5
certifi==2019.9.11
cffi==1.12.3
cfgv==2.0.1
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==4.5.4
cryptography==2.8
daphne==2.4.0
defusedxml==0.6.0
Django==2.2.8
django-cors-headers==3.1.1
django-countries==5.5
django-discover-runner==1.0
django-environ==0.4.5
django-js-asset==1.2.2
django-language-field==0.0.3
django-money==0.15.1
django-mptt==0.11.0
django-phonenumber-field==3.0.1
django-role-permissions==2.2.1
djangorestframework==3.10.3
drf-writable-nested==0.5.1
entrypoints==0.3
environ==1.0
flake8==3.7.8
freezegun==0.3.12
furl==2.1.0
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
identify==1.4.7
idna==2.8
importlib-metadata==0.23
importlib-resources==1.0.2
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
multidict==4.7.5
nodeenv==1.3.3
numpy==1.17.2
oauthlib==3.1.0
orderedmultidict==1.0.1
packaging==19.2
pathspec==0.7.0
pep8==1.7.1
phonenumbers==8.10.20
Pillow==7.0.0
pluggy==0.13.0
pre-commit==1.18.3
psycopg2==2.8.3
py==1.8.0
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycares==3.1.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.2
pytest==5.2.1
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-django==3.5.1
pytest-mock==1.13.0
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
redis==2.10.6
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.5
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
zope.interface==4.7.1
